refactor: 어드민 유저 리팩토링 (#1083)
* refactor: student 패키지 분리

* refactor: student 패키지 분리

* refactor: owner 패키지 분리

* refactor: AdminUserService 리팩토링

* refactor: 테스트 리팩토링

* refactor: 들여쓰기

* fix: 테스트 수정

* fix: getNewOwners 로직 수정

---------
