+++ src/main/java/in/koreatech/koin/admin/student/controller/AdminStudentApi.java
@@ -1,0 +1,81 @@
+package in.koreatech.koin.admin.student.controller;
+
+import static in.koreatech.koin.domain.user.model.UserType.ADMIN;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import in.koreatech.koin.admin.student.dto.AdminStudentResponse;
+import in.koreatech.koin.admin.student.dto.AdminStudentUpdateRequest;
+import in.koreatech.koin.admin.student.dto.AdminStudentUpdateResponse;
+import in.koreatech.koin.admin.student.dto.AdminStudentsResponse;
+import in.koreatech.koin.global.auth.Auth;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+
+@Tag(name = "(Admin) Student: 학생", description = "관리자 권한으로 학생 정보를 관리한다")
+public interface AdminStudentApi {
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
+        }
+    )
+    @Operation(summary = "학생 리스트 조회(페이지네이션)")
+    @SecurityRequirement(name = "Jwt Authentication")
+    @GetMapping("/admin/students")
+    ResponseEntity<AdminStudentsResponse> getStudents(
+        @RequestParam(required = false) Integer page,
+        @RequestParam(required = false) Integer limit,
+        @RequestParam(required = false) Boolean isAuthed,
+        @RequestParam(required = false) String nickname,
+        @RequestParam(required = false) String email,
+        @Auth(permit = {ADMIN}) Integer adminId
+    );
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
+        }
+    )
+    @Operation(summary = "학생 정보 조회")
+    @SecurityRequirement(name = "Jwt Authentication")
+    @GetMapping("/admin/users/student/{id}")
+    ResponseEntity<AdminStudentResponse> getStudent(
+        @PathVariable Integer id,
+        @Auth(permit = {ADMIN}) Integer adminId
+    );
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
+        }
+    )
+    @Operation(summary = "학생 정보 수정")
+    @SecurityRequirement(name = "Jwt Authentication")
+    @PutMapping("/admin/users/student/{id}")
+    ResponseEntity<AdminStudentUpdateResponse> updateStudent(
+        @Valid @RequestBody AdminStudentUpdateRequest adminRequest,
+        @PathVariable Integer id,
+        @Auth(permit = {ADMIN}) Integer adminId
+    );
+}
