+++ src/main/java/in/koreatech/koin/admin/user/controller/AdminUserApi.java
@@ -2,11 +2,9 @@ package in.koreatech.koin.admin.user.controller;
 
 import static in.koreatech.koin.domain.user.model.UserType.ADMIN;
 
-import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -15,24 +13,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import in.koreatech.koin.admin.user.dto.AdminLoginRequest;
 import in.koreatech.koin.admin.user.dto.AdminLoginResponse;
-import in.koreatech.koin.admin.user.dto.AdminNewOwnersResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnerResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnerUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminOwnerUpdateResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnersResponse;
 import in.koreatech.koin.admin.user.dto.AdminPasswordChangeRequest;
 import in.koreatech.koin.admin.user.dto.AdminPermissionUpdateRequest;
 import in.koreatech.koin.admin.user.dto.AdminResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminStudentUpdateResponse;
-import in.koreatech.koin.admin.user.dto.AdminStudentsResponse;
 import in.koreatech.koin.admin.user.dto.AdminTokenRefreshRequest;
 import in.koreatech.koin.admin.user.dto.AdminTokenRefreshResponse;
 import in.koreatech.koin.admin.user.dto.AdminUpdateRequest;
 import in.koreatech.koin.admin.user.dto.AdminsResponse;
 import in.koreatech.koin.admin.user.dto.CreateAdminRequest;
-import in.koreatech.koin.admin.user.dto.OwnersCondition;
 import in.koreatech.koin.admin.user.enums.TeamType;
 import in.koreatech.koin.admin.user.enums.TrackType;
 import in.koreatech.koin.domain.user.model.User;
@@ -48,25 +36,6 @@ import jakarta.validation.Valid;
 
 @Tag(name = "(Admin) User: 회원", description = "관리자 권한으로 회원 정보를 관리한다")
 public interface AdminUserApi {
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "학생 리스트 조회(페이지네이션)")
-    @SecurityRequirement(name = "Jwt Authentication")
-    @GetMapping("/admin/students")
-    ResponseEntity<AdminStudentsResponse> getStudents(
-        @RequestParam(required = false) Integer page,
-        @RequestParam(required = false) Integer limit,
-        @RequestParam(required = false) Boolean isAuthed,
-        @RequestParam(required = false) String nickname,
-        @RequestParam(required = false) String email,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
 
     @ApiResponses(
         value = {
@@ -238,120 +207,6 @@ public interface AdminUserApi {
     ResponseEntity<Void> updateAdminPermission(
         @RequestBody @Valid AdminPermissionUpdateRequest request,
         @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "사장님 권한 요청 허용")
-    @SecurityRequirement(name = "Jwt Authentication")
-    @PutMapping("/admin/owner/{id}/authed")
-    ResponseEntity<Void> allowOwnerPermission(
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "회원 정보 조회")
-    @SecurityRequirement(name = "Jwt Authentication")
-    @GetMapping("/admin/users/student/{id}")
-    ResponseEntity<AdminStudentResponse> getStudent(
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "회원 정보 수정")
-    @SecurityRequirement(name = "Jwt Authentication")
-    @PutMapping("/admin/users/student/{id}")
-    ResponseEntity<AdminStudentUpdateResponse> updateStudent(
-        @Valid @RequestBody AdminStudentUpdateRequest adminRequest,
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 사장님 조회")
-    @SecurityRequirement(name = "Jwt Authentication")
-    @GetMapping("/admin/users/owner/{id}")
-    ResponseEntity<AdminOwnerResponse> getOwner(
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "특정 사장님 수정")
-    @SecurityRequirement(name = "Jwt Authentication")
-    @PutMapping("/admin/users/owner/{id}")
-    ResponseEntity<AdminOwnerUpdateResponse> updateOwner(
-        @PathVariable Integer id,
-        @RequestBody @Valid AdminOwnerUpdateRequest request,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true)))
-        }
-    )
-    @Operation(summary = "가입 신청한 사장님 리스트 조회 (페이지네이션)")
-    @SecurityRequirement(name = "Jwt Authentication")
-    @GetMapping("/admin/users/new-owners")
-    ResponseEntity<AdminNewOwnersResponse> getNewOwners(
-        @ParameterObject @ModelAttribute OwnersCondition ownersCondition,
-        @Auth(permit = {ADMIN}) Integer adminId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true)))
-        }
-    )
-    @Operation(summary = "사장 리스트 조회 (페이지네이션)")
-    @SecurityRequirement(name = "Jwt Authentication")
-    @GetMapping("/admin/users/owners")
-    ResponseEntity<AdminOwnersResponse> getOwners(
-        @ParameterObject @ModelAttribute OwnersCondition ownersCondition,
         @Auth(permit = {ADMIN}) Integer adminId
     );
